Add a (failing) unit test case to reveal the issue
The test case builds an initial tree (/, /foo/bar, /foo/bar/item1)
and checks that the introspection data is correct. Then it adds an
object (/foo/bar/item2) to the tree and checks the introspection data
again.

The test fails because introspect on / lists this new object, but it
shouldn't.